+++ unit-tests/ModelsDynamicOperationsTest.php
@@ -101,20 +101,28 @@ class ModelsDynamicOperationsTest extends PHPUnit_Framework_TestCase
 		);
 
 		$persona = $dynamicPersonasRepository->findFirst();
-		$this->assertTrue($persona->save());
+		$this->assertTrue(
+			$di->get("modelsManager")->save($persona)
+		);
 
 		$this->assertEquals(count($tracer), 3);
 
 		$persona->nombres = 'Other Name '.mt_rand(0, 150000);
+
         $this->assertEquals($persona->getChangedFields(), array('nombres'));
-		$this->assertTrue($persona->save());
+		$this->assertTrue(
+			$di->get("modelsManager")->save($persona)
+		);
 
 		$this->assertEquals('UPDATE `personas` SET `nombres` = ? WHERE `cedula` = ?', $tracer[3]);
 
 		$persona->nombres = 'Other Name '.mt_rand(0, 150000);
 		$persona->direccion = 'Address '.mt_rand(0, 150000);
+
         $this->assertEquals($persona->getChangedFields(), array('nombres', 'direccion'));
-		$this->assertTrue($persona->save());
+		$this->assertTrue(
+			$di->get("modelsManager")->save($persona)
+		);
 
 		$this->assertEquals('UPDATE `personas` SET `nombres` = ?, `direccion` = ? WHERE `cedula` = ?', $tracer[4]);
 	}
@@ -126,20 +134,22 @@ class ModelsDynamicOperationsTest extends PHPUnit_Framework_TestCase
 		);
 
 		$personer = $dynamicPersonersRepository->findFirst();
-		$this->assertTrue($personer->save());
+		$this->assertTrue($di->get("modelsManager")->save($personer));
 
 		$this->assertEquals(count($tracer), 3);
 
 		$personer->navnes = 'Other Name '.mt_rand(0, 150000);
+
         $this->assertEquals($personer->getChangedFields(), array('navnes'));
-		$this->assertTrue($personer->save());
+		$this->assertTrue($di->get("modelsManager")->save($personer));
 
 		$this->assertEquals('UPDATE `personas` SET `nombres` = ? WHERE `cedula` = ?', $tracer[3]);
 
 		$personer->navnes = 'Other Name '.mt_rand(0, 150000);
 		$personer->adresse = 'Address '.mt_rand(0, 150000);
+
         $this->assertEquals($personer->getChangedFields(), array('navnes', 'adresse'));
-		$this->assertTrue($personer->save());
+		$this->assertTrue($di->get("modelsManager")->save($personer));
 
 		$this->assertEquals('UPDATE `personas` SET `nombres` = ?, `direccion` = ? WHERE `cedula` = ?', $tracer[4]);
 	}