+++ unit-tests/ModelsForeignKeysTest.php
@@ -121,7 +121,9 @@ class ModelsForeignKeysTest extends PHPUnit_Framework_TestCase
 		$robotsParts->robots_id = 1;
 		$robotsParts->parts_id = 100;
 
-		$this->assertFalse($robotsParts->save());
+		$this->assertFalse(
+			$di->get("modelsManager")->save($robotsParts)
+		);
 
 		$messages = array(
 			0 => Phalcon\Mvc\Model\Message::__set_state(array(
@@ -136,7 +138,9 @@ class ModelsForeignKeysTest extends PHPUnit_Framework_TestCase
 
 		$robotsParts->robots_id = 100;
 		$robotsParts->parts_id = 1;
-		$this->assertFalse($robotsParts->save());
+		$this->assertFalse(
+			$di->get("modelsManager")->save($robotsParts)
+		);
 
 		$messages = array(
 			0 => Phalcon\Mvc\Model\Message::__set_state(array(
@@ -158,7 +162,9 @@ class ModelsForeignKeysTest extends PHPUnit_Framework_TestCase
 		$robot = $robotsRepository->findFirst();
 		$this->assertNotEquals($robot, false);
 
-		$this->assertFalse($robot->delete());
+		$this->assertFalse(
+			$di->get("modelsManager")->delete($robot)
+		);
 
 		$messages = array(
 			0 => Phalcon\Mvc\Model\Message::__set_state(array(
@@ -178,7 +184,9 @@ class ModelsForeignKeysTest extends PHPUnit_Framework_TestCase
 		$part = $partsRepository->findFirst();
 		$this->assertNotEquals($part, false);
 
-		$this->assertFalse($part->delete());
+		$this->assertFalse(
+			$di->get("modelsManager")->delete($part)
+		);
 
 		$messages = array(
 			0 => Phalcon\Mvc\Model\Message::__set_state(array(
@@ -200,7 +208,9 @@ class ModelsForeignKeysTest extends PHPUnit_Framework_TestCase
 		$robottersDeles = new RobottersDeles();
 		$robottersDeles->robottersCode = 1;
 		$robottersDeles->delesCode = 100;
-		$this->assertFalse($robottersDeles->save());
+		$this->assertFalse(
+			$di->get("modelsManager")->save($robottersDeles)
+		);
 
 		$messages = array(
 			0 => Phalcon\Mvc\Model\Message::__set_state(array(
@@ -215,7 +225,9 @@ class ModelsForeignKeysTest extends PHPUnit_Framework_TestCase
 
 		$robottersDeles->robottersCode = 100;
 		$robottersDeles->delesCode = 1;
-		$this->assertFalse($robottersDeles->save());
+		$this->assertFalse(
+			$di->get("modelsManager")->save($robottersDeles)
+		);
 
 		$messages = array(
 			0 => Phalcon\Mvc\Model\Message::__set_state(array(
@@ -236,7 +248,9 @@ class ModelsForeignKeysTest extends PHPUnit_Framework_TestCase
 		$robotter = $robottersRepository->findFirst();
 		$this->assertNotEquals($robotter, false);
 
-		$this->assertFalse($robotter->delete());
+		$this->assertFalse(
+			$di->get("modelsManager")->delete($robotter)
+		);
 
 		$messages = array(
 			0 => Phalcon\Mvc\Model\Message::__set_state(array(
@@ -256,7 +270,9 @@ class ModelsForeignKeysTest extends PHPUnit_Framework_TestCase
 		$dele = $delesRepository->findFirst();
 		$this->assertNotEquals($dele, false);
 
-		$this->assertFalse($dele->delete());
+		$this->assertFalse(
+			$di->get("modelsManager")->delete($dele)
+		);
 
 		$messages = array(
 			0 => Phalcon\Mvc\Model\Message::__set_state(array(